cephadm: move custom container class to a new file
